+++ preview_docs.sh
@@ -14,7 +14,7 @@ ARG1="${1:-}"
 case "${ARG1}" in
     -h|--help)
         cat << EOF
-Usage:  preview_docs.sh [--rebuild|-h]
+Usage:  preview_docs.sh [--rebuild|-h] [URL_PATH]
 
 Builds a local preview of the documentation in the parent repository that
 includes tk-doc-generator as a sub-module.
@@ -23,6 +23,10 @@ Options:
     --rebuild       Force re-build docker image. By default (for speed),
                     it will only build a new Docker image if one doesn't exist.
     -h, --help      Print this help and exit 0.
+    URL_PATH        Explicit URL sub folder/path to host under.
+                    By default, it's automatically calculated in order:
+                    1. DOC_PATH=/* in .travis.yml
+                    2. Current folder name
 
 Environment variables:
     IMAGE_TAG:
@@ -62,6 +66,7 @@ IMAGE_TAG="${IMAGE_TAG:-tk-doc-generator}"
 if [ -z "$(docker images --quiet ${IMAGE_TAG})" -o "${ARG1}" == "--rebuild" ]
 then
     docker build --tag "${IMAGE_TAG}" ${THIS_DIR}
+    shift  # Process remaining arg as explicit URL_PATH
 else
     echo "'${IMAGE_TAG}' docker image already built"
     echo "To force a re-build, try running this script with --rebuild flag"
@@ -80,12 +85,64 @@ fi
 
 
 # Finally, perform the document generation in container
-docker run --rm \
-    ${MOUNT_FLAGS} \
-    ${IMAGE_TAG} \
-    --url="http://localhost" \
-    --url-path="$(pwd)/_build" \
-    --source="${MOUNT_TO}/docs" \
-    --output="${MOUNT_TO}/_build"
+EXPOSED_PORT="4000"
+URL_PATH="${1:-$(
+    grep -oP '(?<=DOC_PATH=/)[a-zA-Z0-9][a-zA-Z0-9_.-]*' $(pwd)/.travis.yml \
+    || basename $(pwd)
+)}"
 
-echo "Documentation built. Open in web-browser: $(pwd)/_build/index.html"
+if (echo "${URL_PATH}" | grep -qP '^[a-zA-Z0-9][a-zA-Z0-9_.-]*$')
+then
+    CONTAINER_NAME="${URL_PATH}-${EXPOSED_PORT}"
+    until docker run --rm -d \
+            --hostname="${CONTAINER_NAME}" \
+            --name="${CONTAINER_NAME}" \
+            -p "${EXPOSED_PORT}:${EXPOSED_PORT}" \
+            -e EXPOSED_PORT="${EXPOSED_PORT}" \
+            -e URLPATH="/${URL_PATH}" \
+            -e SOURCE="${MOUNT_TO}/docs" \
+            -e OUTPUT="${MOUNT_TO}/_build" \
+            ${MOUNT_FLAGS} \
+            ${IMAGE_TAG}
+    do
+        EXPOSED_PORT="$(( EXPOSED_PORT + 1 ))"
+        CONTAINER_NAME="${URL_PATH}-${EXPOSED_PORT}"
+    done
+else
+    echo "Invalid URL_PATH to host as: '${URL_PATH}'"
+    echo 'It must only contain these characters: [a-zA-Z0-9_.-]'
+    echo
+    echo 'Please either:'
+    echo '1. Set a valid DOC_PATH=/something in $(pwd)/.travis.yml, where that'
+    echo '   something must only contain those valid characters:  e.g.'
+    echo
+    echo "       env:"
+    echo "         global:"
+    echo "             - DOC_PATH=/tk-amazing-tool"
+    echo
+    echo "2. Rename current folder to only contain those valid characters."
+    echo
+    echo "3. Explicitly provide a valid URL_PATH to create under, e.g."
+    echo
+    echo "       preview_docs.sh tk-amazing-tool"
+    echo
+    echo
+    URL_PATH=''
+    while true
+    do
+        CONTAINER_ID=$(docker run --rm -d \
+                -p "${EXPOSED_PORT}:${EXPOSED_PORT}" \
+                -e EXPOSED_PORT="${EXPOSED_PORT}" \
+                -e URLPATH="/" \
+                -e SOURCE="${MOUNT_TO}/docs" \
+                -e OUTPUT="${MOUNT_TO}/_build" \
+                ${MOUNT_FLAGS} \
+                ${IMAGE_TAG}) \
+            && break \
+            || EXPOSED_PORT="$(( EXPOSED_PORT + 1 ))"
+    done
+    CONTAINER_NAME=$(docker ps -f id=${CONTAINER_ID} --format '{{.Names}}')
+fi
+
+# Directly calling log follow here to show any errors in serve_docs.sh
+docker logs --follow ${CONTAINER_NAME}